% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_TopDomData.R
\name{as_TopDomData}
\alias{as_TopDomData}
\title{Coerces \code{hic_bin()} Data into list of TopDomData Objects}
\usage{
as_TopDomData(data)
}
\arguments{
\item{data}{A named list.}
}
\value{
A named list of length \code{length(data)}.
}
\description{
Coerces \code{hic_bin()} Data into list of TopDomData Objects
}
