% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_by_organism.R
\name{compile_by_organism}
\alias{compile_by_organism}
\title{Compile Ramani HiC Data Files into Data Frames stored in RDS files}
\usage{
compile_by_organism(samples, organisms = c("human", "mouse"),
  path = file.path("hicData", "GSE84920"), path_dest = "compiledData")
}
\arguments{
\item{samples}{A character vector of K sample names.}

\item{organisms}{A character vector of N organism names.}

\item{path}{The folder where the input HiC data files
\file{*.{percentages,validPairs}.txt.gz} are located.}

\item{path_dest}{The folder where the compiled RDS files should be saved.}
}
\value{
An K-by-N character matrix of pathnames of RDS files.
Each RDS file contains a data.frame.
}
\description{
Compile Ramani HiC Data Files into Data Frames stored in RDS files
}
\section{Progress updates}{

This function signals \link[progressr:progression]{progressr::progression} updates. To visualize,
or in other ways render, progress information, wrap the call inside a
\link[progressr:with_progress]{progressr::with_progress} call.
}

\section{Parallel processing}{

This function supports processing of (organism, sample):s in parallel
via the \pkg{future} framework.
For example, setting \code{future::plan("multiprocess")} will parallelize
on the local machine.
}

\examples{
\donttest{\dontrun{
progressr::with_progress({
  files <- TopDomStudy::compile_by_organism(
              samples=c("GSM2254215_ML1", "GSM2254219_PL2",
                         "GSM2254216_ML2", "GSM2254218_PL1"),
              organisms="human",
              path="hicData/GSE84920", path_dest="compiledData"
           )
})
print(files)
#                human                                         
# GSM2254215_ML1 "compiledData/GSM2254215_ML1,human,unique.rds"
}}

}
