% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crc32}
\alias{crc32}
\title{CRC32 Checksum of an Object}
\usage{
crc32(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
An eight-character string.
}
\description{
CRC32 Checksum of an Object
}
\keyword{internal}
