% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_score_summary_vs_fraction.R
\name{gg_overlap_score_summary_vs_fraction}
\alias{gg_overlap_score_summary_vs_fraction}
\title{Plots TopDom Overlap Scores and Lengths as a Function of Sample Fraction}
\usage{
gg_overlap_score_summary_vs_fraction(dataset, chromosome, bin_size, rhos,
  window_size = 5L, nsamples = 50L, weights = c("by_length",
  "uniform"), domain_length = NULL, signals = c("mean",
  "test_len_q0.50"), rho_lim = c(0, 1/2), length_lim = c(0, 2e+06),
  rel_heights = c(4, 1), ..., verbose = TRUE)
}
\arguments{
\item{dataset}{(character string) The name of the data set.}

\item{chromosome}{(character string) The chromosome to be plotted.}

\item{bin_size}{(numeric) The bin size (bps) to be plotted.}

\item{rhos}{(numeric vector) The set of fractions (in (0,0.5]) to process.}

\item{window_size}{(integer) The TopDom windows size.
Argument passed as \code{window.size} to \code{\link[TopDom:TopDom]{TopDom::TopDom()}}.}

\item{nsamples}{(integer) Number of random samples for produce.}

\item{weights}{(character string) A character string specifying how overlap
scores across domains should be weighted.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{domain_length}{(optional; character string or numeric vector of length two)
If specified, controls how to filter out too short or too long TopDom domains.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{signals}{(character vector of length two) The two signals to be
plotted in the top and bottom panels.}

\item{rho_lim}{The range of the fraction to be displayed.}

\item{length_lim}{The range of the domain lenghts to be displayed.}

\item{rel_heights}{(numeric vector of length two) The relative height
of the two figures.  Passed to \link[cowplot:plot_grid]{cowplot::plot_grid} as is.}

\item{verbose}{(logical) If \code{TRUE}, verbose output is produced.}

\item{\ldots}{Not used.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plots TopDom Overlap Scores and Lengths as a Function of Sample Fraction
}
\seealso{
Internal, \code{\link[=overlap_score_summary_grid]{overlap_score_summary_grid()}} is used to calculate overlap
scores over (chromosome, bin_size, rho).
}
