% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%<-?\%}
\alias{\%<-?\%}
\title{Assigns a value to a variable, unless it already exists}
\usage{
x \%<-?\% value
}
\arguments{
\item{x}{(RHS) The variables to be assigned.}

\item{value}{(LHS) The expression whose value to assign to the variables.}
}
\value{
(invisibly) the value of the variable.
}
\description{
Assigns a value to a variable, unless it already exists
}
