% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_bin.R
\name{hic_bin}
\alias{hic_bin}
\title{Bins HiC Reads in a Data Frame into Chromosome-Pair Count Matrices}
\usage{
hic_bin(
  data,
  bin_size,
  intra_only = FALSE,
  known_chrs = c(1:22, "X", "Y", "M"),
  progress = TRUE
)
}
\arguments{
\item{data}{An N-by-K data.frame with
characters columns \code{chr_a}, \code{chr_b}, and
numeric columns \code{start_a}, \code{end_a}, \code{start_b}, and \code{end_b}.}

\item{bin_size}{(integer) The bin size (in basepairs).}

\item{intra_only}{(logical) If \code{TRUE}, only intra-chromosome pairs are
binned.}

\item{known_chrs}{(character vector) The names of the L chromosomes
to bin over.}

\item{progress}{(logical) If \code{TRUE}, progress is displayed.}
}
\value{
If \code{intra_only = FALSE}, an L-by-L list matrix where each element \eqn{(i, j)} in turn
holds an integer matrix of binned counts for chromosome pair \eqn{(i, j)}.
If \code{intra_only = TRUE}, an L list where each element \eqn{i} in turn
holds an integer matrix of binned counts for intra-chromosome pair \eqn{(i, i)}.
}
\description{
Bins HiC Reads in a Data Frame into Chromosome-Pair Count Matrices
}
\examples{
pathname <- system.file("compiledData", "human,HAP1,unique,chr=22.rds", package = "TopDomStudy")
reads <- read_rds(pathname)
print(reads)

binned <- hic_bin(reads, bin_size = 100000, intra_only = TRUE)
str(binned)

image(log2(binned[[1]]), axes = FALSE)
box()

}
\author{
Henrik Bengtsson adopted from a script by Adam Olshen who proposed
to use \code{base::table(chr_a, chr_b)} for efficient two-dimensional binning.
}
