% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_topdom_regions.R
\name{import_topdom_regions}
\alias{import_topdom_regions}
\title{Import all test-set TopDom regions into a single data.frame}
\usage{
import_topdom_regions(
 
    pattern = "human,HAP1,unique,bin_size=.*,partition_by=cells_by_half,min_cell_size=2,window_size=.*,test=.*,mainseed=0xBEEF",
  path = "topdomData",
  skip = TRUE,
  save_individual = TRUE
)
}
\arguments{
\item{pattern}{(character string) A regular expression of the subfolders
in folder \code{path} to import.}

\item{path}{(character string) The root folder of the overlap score data.}

\item{skip}{(logical) If TRUE, already existing results are returned, otherwise not.}

\item{save_individual}{(logical) If TRUE, also the imported data.frames of
the subfolders are saved in individual \file{*,test-topdom.rds} file in the
\code{path} folder.}
}
\value{
The pathname of the saved \file{*,test-topdom.rds} file in the
\code{path} folder.
}
\description{
This script will read all available \file{topdomData/*/*.rds}
files, extract the TopDom regions for the test samples (but not the
reference) and convert to a data.frame with additional columns on parameter
settings and RNG seeds appended.  All these data.frames are stacked into one
big data.frame which is saved to a \file{overlapScoreData/*,test-topdom.rds}.
}
