% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_scores_summary.R
\name{overlap_score_summary}
\alias{overlap_score_summary}
\title{Summarizes the TopDom Best Scores across TopDomOverlapScores}
\usage{
overlap_score_summary(fit, drop_reference = TRUE)
}
\arguments{
\item{fit}{A \link[base:list]{base::list} of K \code{TopDomOverlapScores} objects}

\item{drop_reference}{If TRUE, then the reference partition is dropped.}
}
\value{
A \link[base:data.frame]{base::data.frame} with (K-1) rows.
}
\description{
Summarizes the TopDom Best Scores across TopDomOverlapScores
}
