% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds.R
\name{read_rds}
\alias{read_rds}
\title{Robustly Reads an RDS File}
\usage{
read_rds(pathname, ...)
}
\arguments{
\item{pathname}{RDS file to read.}

\item{\ldots}{(optional) Additional arguments passed to \code{\link[base:readRDS]{base::readRDS()}}.}
}
\value{
The \R object read.
}
\description{
Robustly Reads an RDS File
}
\details{
Uses \link[base:readRDS]{base::readRDS} internally but gives a more informative error message
on failure.
}
\keyword{internal}
