% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_overlap_scores.R
\name{read_topdom_domains}
\alias{read_topdom_domains}
\title{Read TopDom Data from Overlap Score Data}
\usage{
read_topdom_domains(pathname, format = c("tibble", "data.frame"))
}
\arguments{
\item{pathname}{(character string) A \file{*,topdom.rds} file.}

\item{format}{(character string) The returned format.}
}
\value{
A data frame with one TopDom domain per row.
}
\description{
Read TopDom Data from Overlap Score Data
}
