% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_overlap_regions.R, R/read_overlap_scores.R
\name{read_topdom_regions}
\alias{read_topdom_regions}
\title{Read TopDom Regions from TopDomStudy Files}
\usage{
read_topdom_regions(pathname, format = c("tibble", "data.frame"))

read_topdom_regions(pathname, format = c("tibble", "data.frame"))
}
\arguments{
\item{pathname}{(character string) A \file{*.rds} file.}

\item{format}{(character string) The returned format.}
}
\value{
A data frame with one TopDom domain per row.

A data frame with TopDom domains with one overlap score and
one length per domain.
}
\description{
Read TopDom Regions from TopDomStudy Files

Read TopDom Regions from TopDomStudy Files
}
