% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_partitions.R
\name{sample_partitions_by_cells_ref_vs_seq}
\alias{sample_partitions_by_cells_ref_vs_seq}
\title{Generate Random, Non-Overlapping Partitions of Cells}
\usage{
sample_partitions_by_cells_ref_vs_seq(reads, ...)
}
\arguments{
\item{reads}{A data.frame of reads.}

\item{\dots}{Argument passed to \link{sample_partitions_similar_weights_ref_vs_seq}.}
}
\value{
A list of random non-overlapping (disjoint) partitions where each
element holds indices in \eqn{{1, 2, ..., n}} and where the union of all
partitions is \eqn{{1, 2, ..., n}}.
}
\description{
Generate Random, Non-Overlapping Partitions of Cells
}
\references{
https://github.com/HenrikBengtsson/SegalM_2017-FISH/issues/16
}
