% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_partitions_by_half.R
\name{sample_partitions_by_half}
\alias{sample_partitions_by_half}
\title{Generate Random, Non-Overlapping Partitions of Indices with Reference vs Rest}
\usage{
sample_partitions_by_half(n, fraction, warn = TRUE)
}
\arguments{
\item{n}{Number of elements to partition.}

\item{fraction}{A numeric in (0,1/2] specifying the size of each partition
relative to \code{n}.}

\item{warn}{If \code{TRUE}, a warning is produced if the partitions produced
are not of equal size.}
}
\value{
A list of disjoint random non-overlapping partitions where each
element holds indices in \eqn{{1, 2, ..., n}}.
The first element holds the \code{reference} partition \eqn{P_0.5} with \eqn{n/2}
indices and the remaining elements holds partitions with indices in
\eqn{{1, 2, ..., n} \ P_0.5} where the size of the partitions corresponds
to the sizes specified by \code{fraction}.
}
\description{
Generate Random, Non-Overlapping Partitions of Indices with Reference vs Rest
}
