% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_partitions.R
\name{sample_partitions_ref_vs_seq}
\alias{sample_partitions_ref_vs_seq}
\title{Generate Random, Non-Overlapping Partitions of Indices with Reference and Sequence of Partitions}
\usage{
sample_partitions_ref_vs_seq(n, fraction = NULL, size = NULL, seq)
}
\arguments{
\item{n}{Number of elements to partition.}

\item{fraction}{A numeric in (0,1] specifying the size of each partition
relative to \code{n}.  If \code{NULL}, argument \code{size} is used.}

\item{size}{An integer in \eqn{{1, 2, ..., n}} specifying the size of each
partition.  If \code{NULL}, argument \code{fraction} is used.}

\item{seq}{A sequence of fractions (if \code{fraction} is specified) or
a sequence of sizes (if \code{size} is specified)}
}
\value{
A list of random non-overlapping partitions where each
element holds indices in \eqn{{1, 2, ..., n}}.
The first element holds the \code{reference} partition \eqn{P_r} with \eqn{|P_r|}
indices and the remaining elements holds partitions with indices in
\eqn{{1, 2, ..., n} \ P_r} where the size of the partitions corresponds
to the sizes specified by \code{seq}.
The reference partition and the "remaining" partitions are always disjoint,
but the "remaining" partitions maybe be overlapping.
}
\description{
Generate Random, Non-Overlapping Partitions of Indices with Reference and Sequence of Partitions
}
