% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_celltype.R
\name{split_by_celltype}
\alias{split_by_celltype}
\title{Split Per-Organism Compiled File Data into Subsets of Cell Types}
\usage{
split_by_celltype(celltypes = list(human = c("HAP1", "HeLa", "GM12878",
  "K562", "Asynchronous", "Nocadazole"), mouse = c("MEF", "Patski")),
  path = "compiledData")
}
\arguments{
\item{celltypes}{A named list of character vectors.}

\item{path}{The folder where input files are located and where output files
are written.}
}
\value{
A named list of the same structure as \code{celltypes} but
where the character string are pathnames to files produced.
}
\description{
Split Per-Organism Compiled File Data into Subsets of Cell Types
}
\details{
This function operates on files produces by \link{compile_by_organism}.
}
\section{Progress updates}{

This function signals \link[progressr:progression]{progressr::progression} updates. To visualize,
or in other ways render, progress information, wrap the call inside a
\link[progressr:with_progress]{progressr::with_progress} call.
}

\examples{
\donttest{\dontrun{
progressr::with_progress({
  files <- TopDomStudy::split_by_celltype(celltypes=list(human="HAP1"),
                                          path="compiledData")
})
print(files)
# $human
#                                 HAP1 
# "compiledData/human,HAP1,unique.rds"
}}

}
