% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_celltype_chromosome.R
\name{split_by_celltype_chromosome}
\alias{split_by_celltype_chromosome}
\title{Split Per-Cell Type Compiled File Data into Per Chromosome}
\usage{
split_by_celltype_chromosome(
  celltypes = list(human = c("HAP1", "HeLa", "GM12878", "K562", "Asynchronous",
    "Nocadazole"), mouse = c("MEF", "Patski")),
  chromosomes = 1:23,
  path = "compiledData"
)
}
\arguments{
\item{celltypes}{A named list of character vectors.}

\item{chromosomes}{Integer vector of chromosomes to be considered.}

\item{path}{The folder where input files are located and where output files
are written.}
}
\value{
A named list of the same structure as \code{celltypes} but
where the character string are pathnames to files produced.
}
\description{
Split Per-Cell Type Compiled File Data into Per Chromosome
}
\details{
This function operates on files produces by \link{split_by_celltype}.
}
\section{Progress updates}{

This function signals \link[progressr:progression]{progressr::progression} updates. To visualize,
or in other ways render, progress information, wrap the call inside a
\link[progressr:with_progress]{progressr::with_progress} call.
}

\examples{
\donttest{\dontrun{
progressr::with_progress({
  files <- TopDomStudy::split_by_celltype_chromosome(
                celltypes=list(human="HAP1"), chromosomes=c(12,16,22),
                path="compiledData")
})
print(files)
# $human
# $human$HAP1
#                                      chr=12 
# "compiledData/human,HAP1,unique,chr=12.rds" 
#                                      chr=16 
# "compiledData/human,HAP1,unique,chr=16.rds" 
#                                      chr=22 
# "compiledData/human,HAP1,unique,chr=22.rds"
}}

}
