% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom_partitions.R
\name{topdom_partitions}
\alias{topdom_partitions}
\title{Fit TopDom Across Partitions}
\usage{
topdom_partitions(
  reads,
  bin_size,
  partition_by,
  rho,
  reference_rho = 1/2,
  nsamples = 100L,
  seed = TRUE,
  chrs = NULL,
  min_cell_size = 1L,
  dataset,
  cell_ids = NULL,
  window_size = 5L,
  path_out = "topdomData",
  mainseed = 48879,
  force = FALSE,
  as = c("pathname", "value"),
  verbose = FALSE
)
}
\arguments{
\item{reads}{A \link[base:data.frame]{base::data.frame}.}

\item{bin_size}{A positive numeric.}

\item{partition_by}{A string specifying how to partition;
one of \code{"reads"}, \code{"cells"}, \code{"reads_by_half"}, or \code{"cells_by_half"}.}

\item{rho, reference_rho}{A numeric in \eqn{(0,1/2]} specifying the relative size of the partitions.}

\item{nsamples}{Number of random samples.}

\item{seed}{Random seed for reproducible (parallel) random number generation (RNG).}

\item{chrs}{(optional, filter) Names of chromosomes to iterate over.
Defaults to the chromosomes in \code{reads$chr_a}.}

\item{min_cell_size}{(optional, filter) The minimum number of reads for a cell to
be included. Cells with less reads are dropped.}

\item{dataset}{(optional) ...}

\item{cell_ids}{(optional, filter) ...}

\item{window_size}{A positive integer passed to \link[TopDom:TopDom]{TopDom::TopDom}.
Defaults to \code{5L}, which is the same as the default in TopDom.}

\item{path_out}{The root folder where to write output.}

\item{mainseed}{...}

\item{force}{If \code{FALSE}, already processed partitions are skipped, otherwise not.}

\item{as}{Should values or pathnames be returned?}

\item{verbose}{If \code{TRUE}, verbose message are produced, otherwise not.}
}
\value{
A named list of length \code{length(chrs)} with names as \code{chrs}.
Each list elements contains \code{nsamples} pathnames of RDS files.
}
\description{
Fit TopDom Across Partitions
}
\section{Parallel processing}{

The future framework is used to parallelize \link[TopDom:TopDom]{TopDom::TopDom} in three layers:
\enumerate{
\item across chromosomes (argument \code{chrs})
\item across random samples (argument \code{nsamples})
\item per random sample, across partitions (argument \code{partition_by})
\itemize{
\item typically two ('reference' and one more)
}
}
}

\examples{
dataset <- "human,HAP1,unique"
pathname <- system.file("compiledData", sprintf("\%s,chr=22.rds", dataset), package = "TopDomStudy")
reads <- read_rds(pathname)
print(reads)

td <- topdom_partitions(reads, bin_size = 100e3, rho = 0.5, nsamples = 5L,
                        partition_by = "cells_by_half",
                        dataset = dataset, verbose = TRUE)

}
