% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Try}
\alias{Try}
\title{Tweaks a Function to be Evaluted using try()}
\usage{
Try(fcn)
}
\arguments{
\item{fcn}{A function.}
}
\value{
A function that calls \code{try(fcn(...), silent = TRUE)}.
}
\description{
Tweaks a Function to be Evaluted using try()
}
\details{
The function  \link[base:stopifnot]{base::stopifnot}.
}
\keyword{internal}
