% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_by.R
\name{unique_by}
\alias{unique_by}
\title{Efficiently Extract Unique Elements from a Data Frame}
\usage{
unique_by(x, by = NA_integer_, ...)
}
\arguments{
\item{x}{A \link[base:data.frame]{base::data.frame}}

\item{by}{The column, as an integer or by name, to be used
to split up \code{x} in chunks.
If \code{NA} (default), then the first factor, string, or
integer column will be used.  If neither type exists, then
the first column will be used.}

\item{\ldots}{Additional arguments passed to \link[base:unique]{base::unique}
per chunk.}
}
\value{
A \link[base:data.frame]{base::data.frame} with duplicated rows drop.
}
\description{
Efficiently Extract Unique Elements from a Data Frame
}
\details{
\code{unique_by(x)} is typically more memory efficient than
\code{unique(x)}.
}
\examples{
iris1 <- unique(datasets::iris)
iris2 <- unique_by(datasets::iris, by = "Species")
stopifnot(identical(iris2, iris1))

}
\keyword{internal}
