% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg_overlap_score_summary_vs_fraction}
\alias{gg_overlap_score_summary_vs_fraction}
\alias{gg_overlap_score_summary_vs_bin_size}
\alias{gg_overlap_score_summary_vs_tad_length}
\title{Plots TopDom Overlap Scores and Lengths as a Function of Sample Fraction or Bin Size}
\usage{
gg_overlap_score_summary_vs_fraction(
  dataset,
  chromosome,
  bin_size,
  rhos,
  window_size = 5L,
  nsamples = 50L,
  weights = c("by_length", "uniform"),
  domain_length = NULL,
  signals = c("mean", "test_len_q0.50"),
  rho_lim = c(0, 1/2),
  length_lim = c(0, 2000),
  rel_heights = c(4, 1),
  line_col = "black",
  fig_path = "figures",
  skip = FALSE,
  ...,
  verbose = TRUE
)

gg_overlap_score_summary_vs_bin_size(
  dataset,
  chromosome,
  bin_sizes,
  rho,
  window_size = 5L,
  nsamples = 50L,
  weights = c("by_length", "uniform"),
  domain_length = NULL,
  signals = c("mean", "test_len_q0.50"),
  bin_size_lim = c(0, max(bin_sizes)),
  length_lim = c(0, 2000),
  rel_heights = c(4, 1),
  line_col = "black",
  fig_path = "figures",
  skip = FALSE,
  ...,
  verbose = TRUE
)

gg_overlap_score_summary_vs_tad_length(
  dataset,
  chromosome,
  bin_sizes,
  rhos,
  window_size = 5L,
  nsamples = 50L,
  weights = c("by_length", "uniform"),
  domain_length = NULL,
  signals = c("mean", "test_len_q0.50"),
  tad_length_lim = c(0, 2e+06),
  line_col = "black",
  fig_path = "figures",
  skip = FALSE,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{(character string) The name of the data set.}

\item{chromosome}{(character string) The chromosome to be plotted.}

\item{bin_size, rho}{(numeric) The bin size (bps) or sample fraction (in (0,0.5]) to be plotted.}

\item{rhos}{(numeric vector) The set of fractions (in (0,0.5]) to process.}

\item{window_size}{(integer) The TopDom windows size.
Argument passed as \code{window.size} to \code{\link[TopDom:TopDom]{TopDom::TopDom()}}.}

\item{nsamples}{(integer) Number of random samples for produce.}

\item{weights}{(character string) A character string specifying how overlap
scores across domains should be weighted.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{domain_length}{(optional; character string or numeric vector of length two)
If specified, controls how to filter out too short or too long TopDom domains.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{signals}{(character vector of length two) The two signals to be
plotted in the top and bottom panels.}

\item{rho_lim, bin_size_lim}{The range of the fraction or bin size to be displayed.}

\item{length_lim}{The range of the domain lenghts to be displayed.}

\item{rel_heights}{(numeric vector of length two) The relative height
of the two figures.  Passed to \link[cowplot:plot_grid]{cowplot::plot_grid} as is.}

\item{line_col}{The color of the profile curve.}

\item{fig_path}{If non-NULL, a PNG image is written to this path.}

\item{skip}{If TRUE and the PNG file already exists, then it is not replotted.}

\item{verbose}{(logical) If \code{TRUE}, verbose output is produced.}

\item{bin_sizes}{(numeric vector) The set of bin sizes (in bps) to process.}

\item{tad_length_lim}{The range of the TAD lengths to be displayed.}

\item{\ldots}{Not used.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plots TopDom Overlap Scores and Lengths as a Function of Sample Fraction or Bin Size
}
\examples{
gg1 <- gg_overlap_score_summary_vs_fraction(
  dataset       = "human,HAP1",
  chromosome    = "22",
  bin_size      = 100*1e3,
  rhos          = c(0.10, 0.20, 0.30, 0.40, 0.50),
  window_size   = 5L,
  weights       = "uniform",
  domain_length = NULL,
  nsamples      = 4L,
  line_col      = "red",
  verbose       = TRUE
)
print(gg1)

gg2 <- gg_overlap_score_summary_vs_bin_size(
  dataset       = "human,HAP1",
  chromosome    = "22",
  bin_sizes     = c(20, 40, 60, 80, 100)*1e3,
  rho           = 0.50,
  window_size   = 5L,
  weights       = "uniform",
  domain_length = NULL,
  nsamples      = 4L,
  line_col      = "blue",
  verbose       = TRUE
)
print(gg2)
}
\seealso{
Internal, \code{\link[=overlap_score_summary_grid]{overlap_score_summary_grid()}} is used to calculate overlap
scores over (chromosome, bin_size, rho).
}
