% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_scores_summary.R
\name{overlap_score_summary}
\alias{overlap_score_summary}
\title{Summarizes the TopDom Best Scores across TopDomOverlapScores}
\usage{
overlap_score_summary(fit, weights = c("uniform", "by_length"),
  domain_length = NULL, drop_reference = TRUE)
}
\arguments{
\item{fit}{A \link[base:list]{base::list} of K \code{TopDomOverlapScores} objects}

\item{weights}{A character string specifying how overlap scores across
domains should be weighted.}

\item{domain_length}{(optional) A length-two numeric vector (min, max)
specifying the minimum and maximum length of domains to include in the
summary.  Any domains with lengths outside this range will be discarded.
If \code{NULL} (default) or \code{c(-Inf, +Inf)}, all domains are included.}

\item{drop_reference}{If \code{TRUE}, then the reference partition is dropped.}
}
\value{
A \link[base:data.frame]{base::data.frame} with (K-1) rows.
}
\description{
Summarizes the TopDom Best Scores across TopDomOverlapScores
}
