% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_score_summary_vs_bin_size.R
\name{overlap_score_summary_vs_bin_size}
\alias{overlap_score_summary_vs_bin_size}
\title{Calculate and Summarize TopDom Overlap Scores as Function of Bin Size}
\usage{
overlap_score_summary_vs_bin_size(dataset, chromosomes, bin_sizes, rhos,
  window_size = 5L, nsamples = 50L, weights = c("by_length",
  "uniform"), domain_length = NULL, verbose = FALSE)
}
\arguments{
\item{dataset}{(character string) The name of the data set.}

\item{chromosomes}{(character vector) Chromosomes to process.}

\item{bin_sizes}{(numeric vector) The set of bin sizes (in bps) to process.}

\item{rhos}{(numeric vector) The set of fractions (in (0,0.5]) to process.}

\item{window_size}{(integer) The TopDom windows size.
Argument passed as \code{window.size} to \code{\link[TopDom:TopDom]{TopDom::TopDom()}}.}

\item{nsamples}{(integer) Number of random samples for produce.}

\item{weights}{(character string) A character string specifying how overlap
scores across domains should be weighted.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{domain_length}{(optional; character string or numeric vector of length two)
If specified, controls how to filter out too short or too long TopDom domains.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{verbose}{(logical) If \code{TRUE}, verbose output is produced.}
}
\value{
A three-dimensional character array of pathname names where the
first dimension specify \code{chromosomes}, the second \code{bin_sizes}, and
the third 'rhos'.
}
\description{
Calculate and Summarize TopDom Overlap Scores as Function of Bin Size
}
\details{
PNG images are written to the \file{figures/} folder (created if missing).
}
\seealso{
Internal, \code{\link[=overlap_score_summary_grid]{overlap_score_summary_grid()}} is used to calculate overlap
scores over (chromosome, bin_size, rho).
}
