% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_score_summary_vs_fraction.R
\name{overlap_score_summary_vs_fraction}
\alias{overlap_score_summary_vs_fraction}
\title{Calculate and Summarize TopDom Overlap Scores as Function of Sample Fraction}
\usage{
overlap_score_summary_vs_fraction(dataset, chromosomes, bin_sizes, rhos,
  window_size = 5L, nsamples = 50L, weights = c("by_length",
  "uniform"), domain_length = NULL, figures = TRUE, verbose = FALSE)
}
\arguments{
\item{dataset}{(character string) The name of the data set.}

\item{chromosomes}{(character vector) Chromosomes to process.}

\item{bin_sizes}{(numeric vector) The set of bin sizes (in bps) to process.}

\item{rhos}{(numeric vector) The set of fractions (in (0,0.5]) to process.}

\item{window_size}{(integer) The TopDom windows size.
Argument passed as \code{window.size} to \code{\link[TopDom:TopDom]{TopDom::TopDom()}}.}

\item{nsamples}{(integer) Number of random samples for produce.}

\item{weights}{(character string) A character string specifying how overlap
scores across domains should be weighted.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{domain_length}{(optional; character string or numeric vector of length two)
If specified, controls how to filter out too short or too long TopDom domains.
Argument passed as is to \code{\link[=overlap_score_summary]{overlap_score_summary()}}.}

\item{figures}{(logical) If TRUE, PNG images are written to the \file{figures/}
folder (created if missing).  This requires the \pkg{ggplot2} packages.}

\item{verbose}{(logical) If TRUE, verbose output is produced.}
}
\value{
A three-dimensional character array of pathname names where the first
dimension specify \code{chromosomes}, the second \code{bin_sizes}, and the third 'rhos'.
}
\description{
Calculate and Summarize TopDom Overlap Scores as Function of Sample Fraction
}
\section{Parallel processing}{

The \pkg{future} framework is used to parallelize \code{\link[TopDom:TopDom]{TopDom::TopDom()}} in 2+3 layers:
\enumerate{
\item across (chromosome, bin_size) (arguments \code{chromosomes} and \code{bin_sizes})
\item across all fractions (argument \code{rhos})
\item \code{\link[=overlap_scores_partitions]{overlap_scores_partitions()}} layers:
a. chromosomes (here a single one)
b. random samples (argument \code{nsamples})
c. partions per sample (typically only two)
}
}

