% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_scores_partitions.R
\name{overlap_scores_partitions}
\alias{overlap_scores_partitions}
\title{Calculates TopDom Overlap Scores Across Partitions}
\usage{
overlap_scores_partitions(reads, dataset, cell_ids, bin_size, partition_by,
  min_cell_size, rho, seed_tags, nsamples = 100L, chrs, path_out = ".",
  save_topdom = TRUE, seed = TRUE, verbose = FALSE)
}
\arguments{
\item{reads}{...}

\item{dataset}{...}

\item{cell_ids}{...}

\item{bin_size}{...}

\item{partition_by}{...}

\item{min_cell_size}{...}

\item{rho}{...}

\item{seed_tags}{...}

\item{nsamples}{= 100L ...}

\item{chrs}{...}

\item{path_out}{...}

\item{save_topdom}{...}

\item{seed}{...}

\item{verbose}{...}
}
\value{
A named list of length \code{length(chrs)}.
}
\description{
Calculates TopDom Overlap Scores Across Partitions
}
