% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_topdom_regions.R
\name{read_topdom_regions}
\alias{read_topdom_regions}
\title{Read TopDom Regions from TopDomStudy Files}
\usage{
read_topdom_regions(pathname, format = c("tibble", "data.frame"))
}
\arguments{
\item{pathname}{(character string) A \file{*.rds} file.}

\item{format}{(character string) The returned format.}
}
\value{
A data frame with one TopDom domain per row.
}
\description{
Read TopDom Regions from TopDomStudy Files
}
