% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_partitions_by_half.R
\name{sample_partitions_similar_weights_by_half}
\alias{sample_partitions_similar_weights_by_half}
\title{Generate Random, Non-Overlapping, Similarly-Weighted, Two-Set Partition of Indices}
\usage{
sample_partitions_similar_weights_by_half(
  w,
  fraction,
  w_tolerance = 0.01,
  max_rejections = 100L,
  warn = TRUE
)
}
\arguments{
\item{w}{Numeric vector of \code{n} non-negative, finite weights.
Weights are normalized such that \code{sum(w)} equals one.}

\item{fraction}{A numeric of length two in (0,1/2] specifying the size of
each partition relative to \code{n}, where the first element specifies the
"reference" partition and the second the "test" partition.
For backward compatible reasons, if of length one, then it's equivalent to
specifying \code{fraction = c(reference = 1/2, test = fraction)}.}

\item{w_tolerance}{Maximum allowed difference between target weight of
each of the two sets (e.g. \code{fraction * 1}) and the actual total weight
of the sets (i.e. \code{sum(w[set])}.  If \emph{both} sets are within the tolerance,
the partition is accepted, otherwise rejected.}

\item{max_rejections}{The maximum number of rejections before giving up.}

\item{warn}{If \code{TRUE}, a warning is produced if the partitions produced
are not of equal size.}
}
\value{
A list of two disjoint random non-overlapping partitions where each
element holds indices in \eqn{{1, 2, ..., n}}.  The first element holds the
'reference' partition and the second the 'test' partition.
Attribute \code{weights} gives the total normalized weight of each set.
If no accepted sample was found, then \code{NA} is returned.
Attribute \code{count} gives the number of internal samples produced before
arriving at an accepted or rejected partition.
}
\description{
Generate Random, Non-Overlapping, Similarly-Weighted, Two-Set Partition of Indices
}
